+++ CHANGES.md
@@ -6,6 +6,10 @@
 
 * Updated default filter language in filter extension's POST search request model to match the extension's documentation [#711](https://github.com/stac-utils/stac-fastapi/issues/711)
 
+### Removed
+
+* Removed the Filter Extension depenency from `AggregationExtensionPostRequest` and `AggregationExtensionGetRequest` [#716](https://github.com/stac-utils/stac-fastapi/pull/716)
+
 ## [3.0.0a3] - 2024-06-13
 
 ### Added